Add support for draft routing
When running services for the "draft" publishing stack, they request
through Plek to talk to the same service, but instead those requests
must be routed to the coresponding draft service if one exists. For
example, when specialist-frontend requests to talk to "content-store",
it must instead talk to the draft-content-store service.

To make this happen, add an option to the
plek-config->environment-variables function to replaces services with
coresponding draft services, where available.

Also as part of this, replace the filtering of service uris to expose to
a particular service, as the previous implementation didn't actually work.