Add support for service-port-aliases to the plek configuration
Currently, the plek configuration is filtered by dependent services,
which has helped to identify where there are missing
dependencies. However, some applications do not use the name of the
service when looking up its location, e.g. rummager is called search,
so these aliases can be used to express that search is an alias for rummager.