Patch rubygems to sort the files and test_files fields on write
Bundler can install gems from git repositories. The contents of the
git repository is downloaded when bundle package is run, and the
gemspec file is loaded, and then overwritten. At this point, the files
and test_files fields are converted to literal arrays.

Depending on what ruby code was evaluated to generate the list of
files and test files, and possibly depending on sources of
randomness (e.g. file ordering on the underlying file system), the
ordering of the resulting array is not reproducible.

While a differing order of files might not have an effect, it does
mean that the hash is different, and this makes hash difference is
inconvenient as bundler is used here as part of a fixed output
derivation, and the lack of reproducibility causes it not to have a
fixed output.