Add support for using the authenticating proxy on the draft stack
Add the option to route requests for the draft-origin to
authenticating-proxy, instead of directly to the router.

Setting the Host header in NGinx was required to get omniauth to work,
and add a special case to the Signon setup to use the draft origin
address as the home and redirect uri for the authenticating proxy.