Switch the development system to use a random passphrase
When creating the Signon Dev user, generate a random passphrase and if
possible store it on the disk. Also make it easy to specify a email or
passphrase for the Dev user by setting environment variables.

Store the generated passphrase on the disk, so that it can be reused
when govuk-system is run again. This makes it possible to save the
password in a password manager for easily logging in to the
development system.