Update the publishing-e2e-tests package
These tests were broken due to packages being updated. The tests for
asset uploading currently fail, but upgrading is a step towards
getting all the tests passing again.

In the future, the package refresh can run the tests to ensure that
upgrades don't break them, but this isn't setup yet.