Update guix to release_6
This includes a change that means that only release tags will be used
for refreshing releases, and not GitHub releases.

This will mean that some apps that use a mixture of release tags and
GitHub releases will get updated, whereas before, they were just being
updated to the latest GitHub release.