Fix and enhance support for Plek
Replace the #:replace-services-with-draft-services-where-available
option with the #:remove-draft-prefix option. The former was
overcomplicated.

The refactoring includes adding aliases in to the environment
variables.

Also changed is the exceptions for the content-stores. Instead of not
using a whitelist, either whitelist non-draft services for the live
content-store, and for the draft content store, whitelist all
services, except those that have a draft equivalent.