Add rudimentary support for handling Elasticsearch data
Currently this just supports importing the govuk index in to
rummager.

This is not very neat, as the way the Elasticsearch data is different
to that of the data for other databases, as there is a file per index
for a given hostname (in contrast to a single file). Also, the
<elasticsearch-connection-config> currently doesn't represent the
index or indicies that are being connected to (whereas the
coresponding records for other databases do).