Remove the signon extension for the tailon service
Signon refuses to start if the schema for the redirect URI is not
https, when being run in the Rails production environment.

This is a bit awkward to configure, which suggests it needs moving
elsewhere, so remove it for now so it doesn't break the Publishing E2E
tests system.