Rework the final component of the govuk-puppet-aws data source
Now instead of looking at each file and trying to match it to an
extract, it looks through the extracts and checks for matching files.

This reversed approach allows for more easily choosing between
alternative files, currently this is supported for gzip'ed and xz
compressed data, with the preference being xz.

xz compression takes quite a bit of time, but for the Publishing API
data dump for example, it can result in a file which is less than half
the size of the gzipped file. This is easier to import in to the
store, and takes up less space once there.