Rework the data-directory-with-index data source
Add additional data to the index, so that the data extracts can be
used directly by this data source, and the govuk-data-standalone
script.

The previous model of using the additional source data doesn't really
make sense any more, as with the current data coming in, there are not
separate source files, so the use of this is currently disabled for
listing extracts.