Update the Publishing E2E Tests
This update brings in some changes that prevent unnecessary use of the
chromedriver-helper gem, along with some fixes for the tests to work
with chromedriver and the new Signon interface. So, once again, some
of the tests are passing.

The useful-packages seem to be required for chromium to start, so add
them back by default once again.