Add `govuk connect`, a new and different way of connecting to things
Connecting GOV.UK things used to be easier, and there were less things to
think about.

Now there are >= 4 conceptual environments, with more on the way. And a couple
of these are split over multiple hosting providers. Sometimes the machines for
a govuk-puppet node class are hosted in one place, sometimes they're in a
couple of places, and in this case, they don't always have the same
functionality, configuration, or applications.

Previously I used some SSH configuration to make logging in to machines, that
got more difficult with the naming in AWS. Long SSH commands involving
govuk_node_list work, and `govukcli ssh` also works.

This script is an evolution of those long commands, running SSH to run
govuk_node_list, and then running SSH with the specific machine information
you get back. While this script is more general than `govukcli ssh`, to just
compare the SSH functionality, `govuk connect ssh`:

 - Doesn't require using an SSH agent, or forwarding it
 - Supports a more universal way of addressing machines
 - Treats environments and hosting providers as separate concepts

To do this, it uses information from the govuk-puppet
hieradata. Unfortunately, at the moment, it's not possible to tell with ease
what's meant to be running where from the hieradata, so data from the
govuk-developer-docs is used to switch between hosting providers.

The `govuk connect` script is also more than a simple SSH wrapper. It can help
you connected to RabbitMQ, Sidekiq Monitoring, and allow jumping directly to
application consoles or application database consoles.

In the future, I can also imagine adding support for connecting to the
govuk-puppet development-vm, `govuk system` containers, and supporting SCP to
allow copying data from and to machines more easily.