Improve the handling the app-* commands
Previously, it was assumed that each app only runs on one class of node. At
least with the current data model, that's not correct, as applications like
the Content Store or Whitehall run on different node types, and the behaviour
on the different node types is significant.

Previously, if you ran:

  govuk connect app-console -e integration whitehall

You would probably end up on a whitehall_frontend node, with little choice
about this.

Now, that same command will error, stating that the node type is ambiguous,
and needs specifying, for example:

  govuk connect app-console -e integration whitehall_backend/whitehall

I don't particularly like using / to separate the optional node class and the
app name, as it's used to separate the hosting provider and node class when
you use other connection types (SSH for example). I think it may make more
sense to use a different symbol, but I couldn't decide which.