Add the govuk-connect repo to the default GOVUK_EXEC_PATH
And move `govuk-guix/bin` itself to the end.

This is part of moving govuk-connect out of govuk-guix. With this
change, it at least works consistently, but the script from the
govuk-connect repository is used.

I'm still not sure things will work in an ideal world, but this is a
useful change for now.